feat: Implement automatic weight estimation and add height tracking
This commit introduces a major new functionality: automatic weight estimation based on steps taken, calories burned, and the user's height. It also adds the ability for users to input and store their height.

- **Automatic Weight & Calorie Estimation:**
    - When navigating to the weight screen, the app now calculates the estimated calories burned and resulting weight loss based on the step count from the previous session.
    - The formula uses step count, user height (to estimate distance), and the last recorded weight.
    - A new "Estimated" weight entry is automatically added to the weight log.
    - The `weight` column in the database and `Weight` model was changed from `INTEGER` to `REAL` to support fractional weight values.

- **Height Tracking and Initial Setup:**
    - Implemented a new `heightDB` to store the user's height.
    - Created a new `input_height.xml` layout for a dialog to capture height.
    - On first launch, the app now prompts the user to enter their height and then their initial weight, ensuring the necessary data for calculations is present.
    - A "Change Height" button was added to the `weightFragment` to allow users to update their height.

- **Code Refinements and Bug Fixes:**
    - Introduced a `Fragments.java` class to store fragment name constants, cleaning up navigation logic in `MainActivity`.
    - Added a `DateFormat.java` class to provide a consistent date format (`dd-MM-yyyy HH:mm:ss`) across the app.
    - The `STEP_COUNT` bundle argument is now correctly removed after being used in `stepsFragment` and `weightFragment` to prevent stale calculations.
    - Changed weight input from `INTEGER` to `double` to allow for more precise entries.
    - In `weightFragment`, added a null check for `getContext()` to prevent crashes.
    - Fixed a bug in `weightAdapter` where `convertView` was not being properly recycled.

+++ app/src/main/java/com/example/trackerst/Weight.java
@@ -9,7 +9,7 @@ public class Weight {
     private String date;
     private String inputType;
 
-    public Weight(int id, @Nullable Double caloriesBurnt, int weight, String date, String inputType) {
+    public Weight(int id, @Nullable Double caloriesBurnt, double weight, String date, String inputType) {
         this.id = id;
         this.caloriesBurnt = caloriesBurnt;
         this.weight = weight;
